Created new file for Channel Options, Unexported channelParams from realtime channel
